pyproject.toml
[project]
name = "langgraph-advance-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
